Added "fetchLisitngs" to FirebaseFirestoreHelper and added fetchListings to ExploreViewController

+++ FBU Final Project/Model/Firestore Models/ItemListing.h
@@ -24,7 +24,7 @@ NS_ASSUME_NONNULL_BEGIN
 @property (nonatomic) NSInteger quantity;
 @property (nonatomic) FIRTimestamp* endsInTimestamp;
 @property (nonatomic, strong) NSString* location;
-@property (nonatomic) CGFloat category;
+@property (nonatomic) NSInteger category;
 
 @end
